requirements.txt
apscheduler>=3.10
discord.py>=2.4
httpx>=0.27
playwright>=1.50
pytrends>=4.9
PyYAML>=6.0
rich>=13.9
